requirements.txt
streamlit
PyMuPDF
sentence_transformers
faiss-cpu